==> code/1-paper-descriptives.Rmd <==
---
title: "1-paper-descriptives"
author: "Johh-Henry Pezzuto"
date: "`r format(Sys.Date(), '%Y-%m-%d')`"
output:
  rmarkdown::html_document:
    smart: true
    theme: spacelab
    number_sections: no
    toc_float:
      collapsed: true
    toc_depth: 4
knitr:
  opts_chunk:
    echo: FALSE
    cache: FALSE
    message: FALSE
    warning: FALSE
    results: 'markup'
editor_options: 
  chunk_output_type: console
---

This document reproduces every quantitative claim and figure in all of the paper's sections up to the meta-analysis itself.

**libraries and data**

```{r setup, message=FALSE}
library(dplyr, warn.conflicts = F)
library(forcats)
library(stringr)
library(tidyr, quietly = T)
library(knitr) 
library(kableExtra)
library(janitor)
library(ggthemes)
library(ggrepel, warn.conflicts = F)
library(ggtext, warn.conflicts = F)
library(patchwork)
library(scales)
library(vtable)

dat <- readRDS(file = '../data/sv_meta_data.rds')

vtable(dat)
```

## 2 Introduction

> Slightly fewer than half (140 of 295) of the studies that we identified for our quantitative meta-analysis measured rape myths in some form

```{r rape_myth_measurement}
dat %>%
  group_by(unique_study_id) %>%
  summarize(has_myth_or_IRMA = any(str_detect(str_to_lower(scale_name), "myth|irma|illinois"))) %>%
  count(has_myth_or_IRMA) |> 
  janitor::adorn_totals(where = "row")
```


## 3 The Problem of Sexual Violence

> Our present review provides empirical results from 13 countries

```{r unique_countries}
length(unique(dat$country))
```

### 4.1 Inclusion and Exclusion Criteria for Study Selection

> (The PRISMA Flow Diagram (Appendix Figure 1) explains how we winnowed the original pool of studies down to the 224 manuscripts)

```{r num_unique_papers}
range(dat$unique_paper_id)
```

> Table 1 details the number of papers for each study design.

```{r table_one}
dat |> 
  distinct(unique_paper_id, unique_study_id, study_design) |> 
  group_by(study_design) |> 
  summarise(n_papers = n_distinct(unique_paper_id),
            n_studies = n_distinct(unique_study_id)) |> 
  kable()
```

## 5 Overview of the Studies in the Meta-Analytic Database

> The database spans the years 1985-2018, encompassing 224 articles that describe 295 studies, in which we coded 489 distinct point estimates.

```{r years_and_countries}
dat |> 
  summarise(year_min = min(year), # 1986, but we start search in 1985
            year_max = max(year),
            n_manuscripts = n_distinct(unique_paper_id),
            n_studies = n_distinct(unique_study_id),
            n_point_estimates = n())

num_data_points <- nrow(dat); num_data_points
```

> Overall these are relatively large studies, with an average of 406 participants and a median of 162 participants. 

```{r avg_study_size}

dat |> 
  group_by(unique_study_id) |> 
  summarise(mean_n_per_study = mean(total_n, na.rm = T)) |>  # collapse outcomes first, so studies with more outcomes don't have more weight
  ungroup() |> 
  summarise(mean_n = mean(mean_n_per_study), # then collapse studies in estimate
            median_n = median(mean_n_per_study))
```

> In addition, 30% percent of these studies test primary prevention efforts that span more than one day.

```{r perc_greater_1_day}
dat |>
  distinct(unique_study_id, intervention_length) |> 
  group_by(unique_study_id) |> 
  filter(intervention_length == max(intervention_length)) |> 
  mutate(long_intervention = ifelse(intervention_length > 1, 1, 0)) |> 
  ungroup() |> 
  summarise(perc_greater_1_day = mean(long_intervention, na.rm = T))
```

> Over these years, we observed an uneven rise in the study of primary prevention efforts. This rise is depicted in Figure 1

```{r studies_over_time}
tt_overall = dat |> # overall line
      group_by(year) |>
      summarise(n = n_distinct(unique_study_id)) |>
      add_row(year = 1985, n = 0)

zeitgeist = tibble(year = c(1998, 2000, 2007), 
                   n = pull(filter(tt_overall, year %in% c(1998, 2000, 2007)), n),
                   event = c("Safe dates", "Men's\nProgram", "Bringing in\nthe Bystander"))

tt_overall     |>
  ggplot(aes(year, n)) +
    geom_label_repel(mapping = aes(x = year, ## labels
                                   y = n,
                                   label = event),
                     size = 6,
                     segment.color = 'gray60',
                     segment.size	= 1.5,
                     inherit.aes = F,
                     data =  zeitgeist,
                     nudge_x = c(2.9, 3.5, 6),
                     nudge_y = c(5.8, 4, 0),
                     min.segment.length = 0, 
                     force = 0.5) +
  scale_x_continuous(
    limits = c(1985, 2018),
    breaks = c(1985, seq(1985, 2020, by = 5), 2018)
  ) +
  geom_line(linewidth = 2) +
  theme_bw() +
  theme(
    axis.text = element_text(size = 12),
    axis.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.position = "bottom"
  ) +
  labs(
    x = NULL,
    y = "# of studies",
    col = NULL
  )

ggsave("../results/figures/tt_studies.pdf", height = 5, width = 7.5)
```

> We observe that 262 studies which constitute 89% of the database were conducted in the United States. 

```{r n_usa}
dat |> 
  distinct(unique_study_id, country) |> 
  count(country, sort = T) |> 
  mutate(perc = round(n / sum(n), 3)) |> 
  filter(country == "USA")
```

> The remaining 33 studies come from various places around the world, including Canada, Europe, and Africa.

```{r country_distribution}
dat |> 
  distinct(unique_study_id, country) |> 
  count(country, sort = T) |> 
  mutate(perc = round(n / sum(n), 3)) |> 
  filter(country != "USA") |> 
  janitor::adorn_totals(where = "row")
```

> Over 66% of all studies are conducted within college campuses. Another 15\% are conducted in high schools, while 11\% are conducted in middle schools. Four percent of studies were conducted in the workplace, and three percent in community settings.

```{r study_settings}
dat |> 
  distinct(unique_study_id, setting, setting2) |> 
  pivot_longer(cols = c("setting", "setting2"), values_to = "setting") |> 
  drop_na(setting) |> 
  count(setting, sort = T) |> 
  mutate(perc = round(n / sum(n), 3))
```

> In fact, 67% of the studies targeted both men and women, while 28\% targeted only men, and 6\% targeted only women.

```{r gender_breakdown}
dat |> # 
  mutate(participant_sex = 
           if_else(participant_sex %in% c("male and female in mixed groups", 
                                          "male and female in separate groups and also in combined groups",
                                          "male and female in separate groups",
                                          "male and female with group composition not specified",
                                          "general population"),
                  "mixed gender groups", 
                  participant_sex)) |>
  distinct(unique_study_id, participant_sex) |>
  count(participant_sex, sort = T) |> 
  mutate(perc = round(n / sum(n), 4)) |> 
  as.data.frame() # to show the decimals in output
```


```{r descriptive_fig}
(country_fig = dat |> 
  distinct(unique_study_id, country) |> 
  mutate(country = ifelse(country == "USA", "USA", "Rest of the world")) |> 
  count(country, sort = T) |> 
  mutate(country = fct_rev(fct_reorder(factor(country), n)),
         perc = n / sum(n)) |>
  ggplot(aes(country, n, label = n, fill = country)) +
  geom_col() +
  theme_bw() +
  theme(axis.text = element_markdown(size = 12), 
        axis.title = element_text(size = 14),
        legend.text = element_text(size = 14),
        plot.title = element_markdown()) +
  scale_y_continuous(breaks = seq(0, 300, by = 100), limits = c(0, 300)) +
  scale_fill_manual(values = ggthemes::solarized_pal()(2)) +
  guides(fill = "none") +
  labs(title = "**a)** Number of Studies by Country",
       x = NULL, 
       y = "# of studies", 
       col = NULL))
  
(setting_fig = dat |> 
  distinct(unique_study_id, setting, setting2) |> 
  pivot_longer(cols = c("setting", "setting2"), values_to = "setting") |> 
  drop_na(setting) |> 
  count(setting, sort = T) |> 
  mutate(setting = fct_rev(fct_reorder(setting, n)), 
         perc = n / sum(n)) |>
  ggplot(aes(setting, n, fill = setting)) +
  geom_col() +
  scale_fill_manual(values = ggthemes::tableau_color_pal('Tableau 10')(6)) +
  labs(title = "**b)** Number of Studies by Setting",
       x = NULL, 
       y = "# of studies", 
       col = NULL) +
  theme_bw() +
    guides(fill = "none") +
  theme(axis.text = element_markdown(size = 12),
        axis.title = element_text(size = 14),
        legend.text = element_text(size = 14),
        plot.title = element_markdown()))

dat_participant_sex <- dat |> # participant sex lines
      mutate(participant_sex = if_else(participant_sex %in% 
                                         c("male and female in mixed groups", 
                                           "male and female in separate groups and also in combined groups",
                                           "male and female in separate groups",
                                           "male and female with group composition not specified",
                                           "general population"), 
                                      "mixed gender groups", 
                                      participant_sex)) |>
      distinct(year, unique_study_id, participant_sex) |>
      filter(participant_sex %in% c("mixed gender groups", "only male", "only female")) |>
      count(year, participant_sex) |>
      add_row(year = 1985, participant_sex = "mixed gender groups") |>
      complete(year, participant_sex, fill = list(n = 0)) |>
      mutate(participant_sex = str_to_title(participant_sex)) |>
  mutate(participant_sex = factor(participant_sex, levels = c("Mixed Gender Groups", "Only Male", "Only Female")))

(tt_gender_fig = dat_participant_sex |>
  ggplot(aes(year, n, col = participant_sex)) +
  geom_line(linewidth = 2) +
  scale_x_continuous(
    limits = c(1985, 2018),
    breaks = c(1985, seq(1985, 2020, by = 5), 2018)
  ) +
  scale_color_manual(values = ggthemes::palette_pander(3)) +
  theme_bw() +
  theme(
    axis.text = element_markdown(size = 12), 
    axis.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    plot.title = element_markdown(),
    legend.position = "bottom"
  ) +
  labs(
    title = "**c)** Number of Studies Over Time by Group Gender (1985-2018)",
    x = NULL,
    y = "# of studies",
    col = NULL
  ))

country_fig / setting_fig / tt_gender_fig

ggsave("../results/figures/descriptive.pdf", width = 7, height = 7)
```


## 6 The Leading Approach to Reducing Sexual Violence

> As depicted in Figure 3, in our database, ideas about sexual violence are consistently measured more often than behaviors (typically self-reported).

```{r fig_tt_att_beh}
dat |> 
  mutate(scale_type = str_to_title(ifelse(scale_type == "ideas", "ideas", "behavior"))) |> 
  distinct(year, unique_study_id, scale_type) |> 
  count(year, scale_type) |> 
  add_row(year = 1985, scale_type = "Behavior") |> 
  complete(year, scale_type, fill = list(n = 0)) |> 
  mutate(scale_type = factor(scale_type, levels = c("Ideas", "Behavior"))) |> 
  ggplot(aes(x = year, y = n, color = scale_type)) +
  geom_line(linewidth = 3) + 
  labs(
    x = NULL,
    y = "# of studies",
    col = NULL
  ) +
  scale_x_continuous(
    limits = c(1985, 2018),
    breaks = c(1985, seq(1985, 2020, by = 5), 2018)
  ) +
  scale_color_manual(values = ggthemes::tableau_color_pal('Color Blind')(2)) +
  theme_bw() +
  theme(
    axis.text = element_text(size = 12), 
    axis.title = element_text(size = 14),
    legend.text = element_text(size = 14),
    legend.position = "bottom"
  )

ggsave("../results/figures/tt_att_beh.pdf", height = 5)
```

#### 6.1.1 The First ten years

> In the first 10 years, we find 44 studies reported in 32 articles
 
```{r early_years_study_count}
dat |> 
  filter(between(year, 1985, 1995)) |> 
  summarise(n_manuscript = n_distinct(unique_paper_id),
            n_studies = n_distinct(unique_study_id))
dat |> 
  mutate(early = between(year, 1985, 1995)) |> 
  distinct(unique_study_id, .keep_all = T) |> 
  summarise(perc_early = mean(early))
```

> This is also apparent when looking at the type of publication; while 24 of these studies are from journal articles, 20 are from dissertations

```{r early_years_published_dissertation_breakdown}
dat |> 
  filter(between(year, 1985, 1995)) |> 
  distinct(unique_study_id, publication_type) |> 
  count(publication_type)
```

> The majority of these studies are laboratory studies (n = 27) as opposed to field studies (n = 20) that encompass the larger programmatic evaluations that dominate the next generation of studies.

```{r early_years_lab_field} 
dat |> 
  filter(between(year, 1985, 1995)) |> 
  distinct(unique_study_id, lab_field, publication_type) |> 
  count(lab_field)
```

> These studies tend to be smaller than the next generation of studies (with an average of 200 participants)

```{r early_years_study_size}
dat |> 
  filter(between(year, 1985, 1995)) |> 
  group_by(unique_study_id) |> 
  summarise(mean_n_per_study = mean(total_n, na.rm = T)) |>  # collapse outcomes first, so studies with more outcomes don't have more weight
  ungroup() |> 
  summarise(mean_n = mean(mean_n_per_study)) # then collapse studies in estimate
```

> They target both men and women (31 interventions, which constitute 67\% of all studies in this time frame), but a relatively high number of studies target only men (11 studies, or 24\% of all studies in this time frame). 

```{r early_years_gender_breakdown}
dat |> 
  filter(between(year, 1985, 1995)) |> 
      mutate(participant_sex = if_else(participant_sex %in% c("male and female in mixed groups", 
                                         "male and female in separate groups and also in combined groups",
                                          "male and female in separate groups",
                                         "male and female with group composition not specified",
                                         "general population"), 
                                      "mixed gender groups", 
                                      participant_sex)) |>
      distinct(unique_study_id, participant_sex) |>
  count(participant_sex, sort = T) |> 
  mutate(perc = n / sum(n))

#' Note that the Ns don't quite match up here -- some studies have multiple 
#' treatment arms whe (some with mixed gender groups, some with single-gender)
#' so they get reported twice when we tally participant_sex up
```

## 7 Trends in Measurement in the Database

> Overall, outcomes referring to an individual's ideas about sexual violence outnumber behavioral outcomes in our sample by 356 to 143. 

```{r ideas_vs_behavior_outcome_count}
dat |> 
  select(scale_type) |> 
  group_by(scale_type) |> 
  summarise(n = n())
```

> 199 measure outcomes about ideas only

```{r att_only}
dat |> 
  select(unique_study_id, scale_type) |> 
  group_by(unique_study_id) |> 
  summarise(perc_ideas = sum(scale_type == "ideas") / n()) |> 
  filter(perc_ideas == 1) |> 
  ungroup() |> 
  nrow()
```

> 34 measure only behaviors; 

```{r behaviors_only}
dat |> 
  select(unique_study_id, scale_type) |> 
  group_by(unique_study_id) |> 
  summarise(perc_ideas = sum(scale_type == "ideas") / n()) |> 
  filter(perc_ideas == 0) |> 
  ungroup() |> 
  nrow()
```

> 62 measure both

```{r}
dat |> 
  select(unique_study_id, scale_type) |> 
  group_by(unique_study_id) |> 
  summarise(perc_ideas = sum(scale_type == "ideas") / n()) |> 
  filter(!perc_ideas %in% c(0,1)) |> 
  ungroup() |> 
  nrow()
```

> While in the first 15 years, the ratio between ideas-based outcomes and behavior outcomes was 1:4.59, 

```{r early_years_ratio_between_ideas_and_beahvior_outcomes}
dat |> 
  filter(between(year, 1985, 2000)) |>
  count(scale_type) |> 
  pivot_wider(names_from = scale_type, values_from = n) |> 
  mutate(ratio = ideas/behavior)
```

> after 2010 this ratio drops to 1:1.35. 

```{r post_2010_ratio_ideas_behavvior}
dat |> 
  filter(year > 2010) |>
  count(scale_type) |> 
  pivot_wider(names_from = scale_type, values_from = n) |> 
  mutate(ratio = ideas/behavior) |> 
  as.data.frame()
```

### 7.1. Measuring Sexually Violent Behavior

> behavioral outcomes...fall into four categories: perpetration, victimization, bystander, and involvement. Figure 4 demonstrates the trends over time in these measurements.

```{r behavioral_outcomes_by_decade}
dat |>
  filter(behavior_type != "ideas") |> 
  distinct(unique_study_id, decade, behavior_type) |> 
  count(decade, behavior_type) |> 
  complete(decade, behavior_type, fill = list(n = 0)) |> 
  mutate(behavior_type = factor(behavior_type, levels = c("Perpetration", "Victimization", "Bystander", "Involvement"))) |> 
  ggplot(aes(x = decade, y = n, color = behavior_type, group = behavior_type)) +
  geom_line(linewidth = 3) + 
  labs(x = NULL, 
       y = "# of studies", 
       col = NULL) +
  scale_color_manual(values = ggthemes::stata_pal("s1color")(4)) +
  theme_bw() +
    theme(
    axis.text = element_text(size = 12),
    axis.title = element_text(size = 14),
    legend.text = element_text(size = 14),    
    legend.position = "bottom",
    legend.box = "vertical"
  )

ggsave("../results/figures/tt_beh_type.pdf")
```

### 7.2 Measuring ideas

> Overall these and other myth measures constitute 57% of the measures we coded that assessed ideas about sexual violence

```{r myth_percentage}
dat |> 
  select(scale_name, behavior_type) |> 
  filter(behavior_type == "ideas") |> 
  mutate(myth_scale = str_detect(str_to_lower(scale_name), "myth|irma|illinois")) |> 
  group_by(myth_scale) |> 
  summarise(n = n()) |> 
  mutate(perc = n / sum(n))
```
