==> code/5-paper-descriptives.Rmd <==
---
title: "5-descriptives"
author: "Seth Green and Johh-Henry Pezzuto"
date: "`r format(Sys.Date(), '%Y-%m-%d')`"
output:
  rmarkdown::html_document:
    smart: true
    theme: spacelab
    number_sections: no
    toc_float:
      collapsed: true
    toc_depth: 4
editor_options: 
  chunk_output_type: console
---


```{r echo=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      cache = FALSE,
                      message = FALSE,
                      warning = FALSE, 
                      results = 'markup')


library(tidyverse)
library(ggtext)
library(patchwork)
library(ggthemes)
library(ggrepel)
library(knitr) 
library(kableExtra)
library(janitor)
library(scales)
library(metafor)

source('./functions/study_count.R')
source('./functions/map_robust.R')

dat <- readRDS(file = '../data/sa_meta_data_final.rds') %>%
  mutate(
    attitudes_behaviors = case_when(
      all(c('attitudes', 'behavior') %in% scale_type) ~ 0,
      scale_type == 'attitudes' ~ 1,
      scale_type == 'behavior' ~ 2)) %>%
  ungroup()
```

# The Problem of Sexual Violence

Our present review provides empirical results from 13 countries

```{r}
length(unique(dat$country))
```


## Inclusion and Exclusion Criteria for Study Selection

Table 1 details the number of papers for each study design.

```{r}
dat |> 
  distinct(unique_paper_id, unique_study_id, study_design) |> 
  group_by(study_design) |> 
  summarise(n_papers = n_distinct(unique_paper_id),
            n_studies = n_distinct(unique_study_id)) |> 
  kable(format = "latex", booktabs = T)
```


# Overview of the Studies in the Meta-Analytic Database

The database
spans the years 1985-2018, encompassing 226 manuscripts that describe 304 studies, in which we coded 500
distinct point estimates.

```{r years_and_countries}
dat |> 
  summarise(year_min = min(year),
            year_max = max(year),
            n_manuscripts = n_distinct(unique_paper_id),
            n_studies = n_distinct(unique_study_id),
            n_point_estimates = n())
```


Overall these are relatively large studies, with an average of 403 participants and a median of 162 participants. 

```{r}
dat |> 
  group_by(unique_study_id) |> 
  summarise(mean_n_per_study = mean(n_t_post, na.rm = T) + mean(n_c_post, na.rm = T)) |>  # collapse outcomes first, so studies with more outcomes don't have more weight
  ungroup() |> 
  summarise(mean_n = mean(mean_n_per_study), # then collapse studies in estimate
            median_n = median(mean_n_per_study))
```



In addition, 30% percent of these studies test primary prevention efforts that span more than one day.

```{r}
dat |>
  distinct(unique_study_id, intervention_length) |> 
  mutate(intervention_length = ifelse(intervention_length > 1, 1, 0)) |> 
  summarise(perc_greater_1_day = mean(intervention_length, na.rm = T))
```


```{r}
tt_overall = dat |> # overall line
      group_by(year) |>
      summarise(n = n_distinct(unique_study_id)) |>
      add_row(year = 1985, n = 0)

zeitgeist = tibble(year = c(1998, 2000, 2007), 
                   n = pull(filter(tt_overall, year %in% c(1998, 2000, 2007)), n),
                   event = c("Safe dates", "Men's Program", "Bringing in\nthe Bystander"))

tt_overall     |>
  ggplot(aes(year, n)) +
  geom_line(size = 2) +
  # geom_segment(aes(x = year, ## add "flag poles"
  #                  xend = year,
  #                  y = 0,
  #                  yend = n,
  #                  label = NULL),
  #              inherit.aes = F, 
  #              linewidth = 1.5,
  #              data = zeitgeist) +
    geom_label_repel(mapping = aes(x = year, ## labels
                                   y = n,
                                   label = event),
                     size = 6,
                     segment.size	= 1.5,
                     inherit.aes = F,
                     data =  zeitgeist,
                     nudge_x = c(2, 2.5, -1.9),
                     nudge_y = c(6, 5.5, 5),
                     min.segment.length = 0, 
                     force = 0.3) +
  scale_x_continuous(
    limits = c(1985, 2018),
    breaks = c(1985, seq(1985, 2020, by = 5), 2018)
  ) +
  theme_bw() +
  theme(
    axis.text = element_text(size = 12), 
    legend.text = element_text(size = 12),
    legend.position = "bottom"
  ) +
  labs(
    x = NULL,
    y = "# of studies",
    col = NULL
  )

ggsave("results/tt_studies.pdf", height = 5)
```

We observe that 270 studies which constitute 89% of the database were conducted in the United States. The remaining 34 studies
come from various places around the world, including Canada, Europe, and Africa.


```{r}
dat |> 
  distinct(unique_study_id, country) |> 
  count(country, sort = T) |> 
  mutate(perc = n / sum(n))
```


Nearly 69% of all studies are conducted among college students. Another 13% focus on high school students, while 9% percent
focus on middle school students. Outside of the education sphere, we find that 4% of studies were conducted
in the workplace, and 2% in community settings. 

```{r}
dat |> 
  distinct(unique_study_id, setting) |> 
  count(setting, sort = T) |> 
  mutate(perc = n / sum(n))
```


In fact, 61% of the studies targeted both men and women, while 27% targeted only men, and 6% targeted only women (1% of studies did not report the gender target).
```{r}
dat |> # 
      mutate(participant_sex = ifelse(participant_sex %in% c("male and female in mixed groups", 
                                                             "male and female with group composition not specified", 
                                                             "male and female in separate groups"), 
                                      "mixed gender groups", 
                                      participant_sex)) |>
      distinct(unique_study_id, participant_sex) |>
  count(participant_sex, sort = T) |> 
  mutate(perc = n / sum(n))
```


```{r descriptive_fig}
(country_fig = dat |> 
  distinct(unique_study_id, country) |> 
  mutate(country = ifelse(country == "USA", "USA", "Rest of the world")) |> 
  count(country, sort = T) |> 
  mutate(country = fct_rev(fct_reorder(factor(country), n)),
         perc = n / sum(n)) |>
  ggplot(aes(country, n, label = n, fill = country)) +
  geom_col() +
  theme_bw() +
  theme(axis.text = element_markdown(size = 12), 
        plot.title = element_markdown()) +
  scale_y_continuous(breaks = seq(0, 300, by = 100), limits = c(0, 300)) +
   scale_fill_manual(values = ggthemes::solarized_pal()(2)) +
  guides(fill = "none") +
  labs(title = "**a)** Number of Studies by Country",
       x = NULL, 
       y = "# of studies", 
       col = NULL))
  
(setting_fig = dat |> 
  distinct(unique_study_id, .keep_all = T) |> 
  count(setting, sort = T) |> 
  drop_na(setting) |> 
  mutate(perc = n / sum(n)) |>
  ggplot(aes(setting, n, fill = setting)) +
  geom_col() +
  scale_fill_manual(values = ggthemes::tableau_color_pal('Tableau 10')(6)) +
  labs(title = "**b)** Number of Studies by Setting",
       x = NULL, 
       y = "# of studies", 
       col = NULL) +
  theme_bw() +
    guides(fill = "none") +
  theme(axis.text = element_markdown(size = 12),
        plot.title = element_markdown()))


dat_participant_sex <- dat |> # participant sex lines
      mutate(participant_sex = ifelse(participant_sex %in% c("male and female in mixed groups", 
                                                             "male and female with group composition not specified", 
                                                             "male and female in separate groups"), 
                                      "mixed gender groups", 
                                      participant_sex)) |>
      distinct(unique_study_id, .keep_all = T) |>
      filter(participant_sex %in% c("mixed gender groups", "only male", "only female")) |>
      count(year, participant_sex) |>
      add_row(year = 1985, participant_sex = "mixed gender groups") |>
      complete(year, participant_sex, fill = list(n = 0)) |>
      mutate(participant_sex = str_to_title(participant_sex)) |>
  mutate(participant_sex = factor(participant_sex, levels = c("Mixed Gender Groups", "Only Male", "Only Female")))



(tt_gender_fig = dat_participant_sex |>
  ggplot(aes(year, n, col = participant_sex)) +
  geom_line(size = 2) +
  scale_x_continuous(
    limits = c(1985, 2018),
    breaks = c(1985, seq(1985, 2020, by = 5), 2018)
  ) +
  scale_color_manual(values = ggthemes::palette_pander(3)) +
  theme_bw() +
  theme(
    axis.text = element_markdown(size = 12), 
    legend.text = element_markdown(size = 12),
    plot.title = element_markdown(),
    legend.position = "bottom"
  ) +
  labs(
    title = "**c)** Number of Studies Over Time by Group Gender (1985-2018)",
    x = NULL,
    y = "# of studies",
    col = NULL
  ))

country_fig / setting_fig / tt_gender_fig

ggsave("results/descriptive.pdf", width = 7, height = 7)
```


## Zeitgeist Programs

### The first 15 Years

In the first 15 years, we find 110 studies reported in 77 manuscripts (constituting 36% of our database). 
```{r}
dat |> 
  filter(between(year, 1985, 2000)) |> 
  summarise(n_manuscript = n_distinct(unique_paper_id),
            n_studies = n_distinct(unique_study_id))



dat |> 
  mutate(early = between(year, 1985, 2000)) |> 
  distinct(unique_study_id, .keep_all = T) |> 
  summarise(perc_early = mean(early))
```



This is also apparent when looking at the type of publication; while 56 of these studies are journal articles, 54 are dissertations
```{r}
dat |> 
  filter(between(year, 1985, 2000)) |> 
  distinct(unique_study_id, publication_type) |> 
  count(publication_type)
```



The majority of these studies are laboratory studies (n = 61) as opposed to field studies (n = 49) that encompass the larger programmatic evaluations that dominate the next generation of studies.
```{r}
dat |> 
  filter(between(year, 1985, 2000)) |> 
  distinct(unique_study_id, lab_field, publication_type) |> 
  count(lab_field)
```


These studies tend to be smaller than the next generation of studies (with an average of 188 participants)

```{r}
dat |> 
  filter(between(year, 1985, 2000)) |> 
  group_by(unique_study_id) |> 
  summarise(mean_n_per_study = mean(n_t_post, na.rm = T) + mean(n_c_post, na.rm = T)) |>  # collapse outcomes first, so studies with more outcomes don't have more weight
  ungroup() |> 
  summarise(mean_n = mean(mean_n_per_study)) # then collapse studies in estimate
```


which constitute 56\% of all studies in this time frame), but a relatively high number of studies that target only men (32 studies, which constitute 29\% of all studies in this time frame).
```{r}
dat |> 
  filter(between(year, 1985, 2000)) |> 
      mutate(participant_sex = ifelse(participant_sex %in% c("male and female in mixed groups", 
                                                             "male and female with group composition not specified", 
                                                             "male and female in separate groups"), 
                                      "mixed gender groups", 
                                      participant_sex)) |>
      distinct(unique_study_id, participant_sex) |>
  count(participant_sex, sort = T) |> 
  mutate(perc = n / sum(n))
```

## Trends in Measurement in the Database

Overall, mentalizing outcomes outnumber behavioral outcomes in our sample by 355 to 145. 207 studies
measure mentalizing outcomes only; 36 measure only behaviors; and 61 measure both
```{r}

```


While in the first 15 years, the ratio between mentalizing outcomes and behavior outcomes was
roughly 1 to 4, after 2010 this ratio drops to 1 to 1.5.
```{r}

```


Fig XXX displays the relationship between these outcomes over time.

```{r fig_att_beh_corr}
dat %>% 
  mutate(scale_type = str_to_title(ifelse(scale_type == "attitudes", "mentalizing", "behavior"))) |> 
  distinct(year, unique_study_id, scale_type) |> 
  count(year, scale_type) |> 
  add_row(year = 1985, scale_type = "Behavior") |> 
  complete(year, scale_type, fill = list(n = 0)) |> 
  ggplot(aes(x = year, y = n, color = scale_type)) +
  geom_line() + 
  labs(
    x = NULL,
    y = "# of studies",
    col = NULL
  ) +
  scale_x_continuous(
    limits = c(1985, 2018),
    breaks = c(1985, seq(1985, 2020, by = 5), 2018)
  ) +
  theme_bw() +
  theme(
    axis.text = element_text(size = 12), 
    legend.text = element_text(size = 12),
    legend.position = "bottom"
  )

ggsave("output/fig_att_beh_corr.pdf")
```





```{r}
dat |> 
  select(scale_name, behavior_type) |> 
  filter(behavior_type == "Attitude") |> 
  mutate(myth_scale = str_detect(str_to_lower(scale_name), "myth")) |> 
  group_by(myth_scale) |> 
  summarise(n = n())

```




```{r}
dat |> 
  distinct(unique_study_id, .keep_all = T) |> 
  count(setting, sort = T) |> 
  mutate(perc = n / sum(n)) |> 
  janitor::adorn_totals()
```

```{r}
dat |> 
  distinct(unique_paper_id, .keep_all = T) |> 
  count(setting, sort = T) |> 
  mutate(perc = n / sum(n)) |> 
  janitor::adorn_totals()
```

```{r}
dat |> 
    mutate(participant_sex = ifelse(participant_sex %in% c("male and female in mixed groups", 
                                                           "male and female with group composition not specified", 
                                                           "male and female in separate groups"), 
                                    "mixed gender groups",
                                    participant_sex)) |> 
  distinct(unique_paper_id, .keep_all = T) |> 
  count(participant_sex, sort = T) |> 
  mutate(perc = n / sum(n)) |> 
  janitor::adorn_totals()
```








```{r}
dat %>%
  drop_na(study_design) |> 
  ggplot(aes(x = se_d, y = d)) +
  geom_point(aes(
      color = behavior_type,
      shape = study_design),
    size = 3
  ) +
  geom_smooth(
    method = "lm",
    se = F,
    size = 2
  ) +
  #ggtitle("Relationship between effect sizes and standard errors") +
  labs(
    x = "Standard errors",
    y = "Effect size",
    color = "Outcome type",
    shape = "Study Design"
  ) +
  theme_bw() +
  theme(
    axis.text = element_text(size = 12),
    legend.text = element_text(size = 12),
    legend.position = "bottom",
    legend.box = "vertical"
  )

ggsave("../results/d_se.pdf", height = 5)

```


```{r}
dat %>% 
  filter(has_both == 'both') %>% 
  drop_na(study_design) |> 
  group_by(author, year, study_design, unique_study_id, scale_type) %>%
  #filter(delay == min(delay)) |> 
  summarise(mean_d = mean(d), 
         mean_var_d = mean(var_d), 
         mean_se_d = mean(se_d)) %>% 
  pivot_wider(id_cols = c(author, year, unique_study_id, study_design),
    names_from = c(scale_type), 
    values_from = c(mean_d, mean_var_d, mean_se_d)) %>%
  ggplot(aes(x = mean_d_attitudes,
             y = mean_d_behavior)) + 
  geom_point(aes(color = study_design),
             size = 3) + 
  geom_abline(slope = 1, 
              lty = 'dashed',
              size = 2) +
  geom_smooth(lty = "dashed", 
              method = "lm", 
              se = FALSE,
              color = 'grey',
              size = 2) +  
  scale_x_continuous(breaks = seq(-.5, 1.5, by = .5), limits = c(-.5,1.6)) +
  scale_y_continuous(breaks = seq(-.5, 1.5, by = .5), limits = c(-.5,1.6)) +
  labs(x = "Effect size (attitudes)",
       y = "Effect size (behaviors)",
       color = "Study Design") +
  guides(color = guide_legend(title = "Study Design")) + 
  theme_bw() +
    theme(
    axis.text = element_text(size = 12),
    legend.text = element_text(size = 12),
    legend.position = "bottom",
    legend.box = "vertical"
  )

ggsave("../results/att_beh_corr.pdf", width = 7, height = 7)
```


```{r}
dat %>%
  filter(behavior_type != "Attitude") |> 
  distinct(unique_study_id, year, behavior_type) |> 
  count(year, behavior_type) |> 
  add_row(year = 1985:1990, behavior_type = "Involvement") |> # start at 0 for all behaviors. here i use arbitrary behavior 
  complete(year, behavior_type, fill = list(n = 0)) |> 
  ggplot(aes(x = year, y = n, color = behavior_type)) +
  geom_line(size = 3) + 
  labs(title = "**a)** Number of Studies by Country",
       x = NULL, 
       y = "# of studies", 
       col = NULL) +
  scale_x_continuous(
    limits = c(1985, 2018),
    breaks = c(1985, seq(1985, 2020, by = 5), 2018)
  ) +
  theme_bw() +
    theme(
    axis.text = element_text(size = 12),
    legend.text = element_text(size = 12),
    legend.position = "bottom",
    legend.box = "vertical"
  )

```


Table XX details the distribution of behavioral outcomes within studies that employ observational designs (i.e., quasi-experimental and observational studies) and randomized designs (i.e., experiments and horse-race experiments). 
```{r}

dat |> 
  filter(scale_type == "behavior") |> 
  distinct(unique_study_id, study_design, behavior_type) |> 
  count(study_design, behavior_type) |> 
  pivot_wider(values_from = n, id_cols = behavior_type, names_from = study_design, values_fill = 0) |> 
  janitor::adorn_totals() |> 
  kable(format = "latex", booktabs = T)
```


We recovered XXX manuscripts and XXX studies from our search that qualified methodologically and theoretically.

```{r}
dat %>% 
  select(unique_study_id, unique_paper_id) %>% 
  summarise(n_papers = n_distinct(unique_paper_id),
            n_studies = n_distinct(unique_study_id))
```



### Overview of Studies in the Meta-Analytic Database

```{r }
dat |> 
  select(contains("n_")) |> 
  View()

# average N?
mean(dat$n_t_post) + mean(dat$n_c_post)
mean(dat$n_t_pre, na.rm = T) + mean(dat$n_c_pre, na.rm = T)

# participant sex
table(dat$condition_gender)

## intervention length
mean(dat$intervention_length, na.rm = T)
quantile(dat$intervention_length, probs = seq(0, 1, by = 0.2), na.rm = TRUE)
```

### Meta-Analytic Results

(see 3-sa-meta.Rmd)


We found XX studies that measured self-reported behavior.
```{r}
dat %>% 
  select(unique_study_id, scale_type) %>% 
  group_by(scale_type) %>% 
  summarise(n_unique_studies = n_distinct(unique_study_id)) 
```

XX of these studies measured behavior only in the immediate aftermath of the intervention, while xx included long-term measures.  
```{r}
dat %>% 
  filter(scale_type == "behavior") %>% 
  select(unique_study_id, delay) %>% 
  drop_na(delay) %>% 
  group_by(unique_study_id) %>% 
  mutate(max_delay = max(delay),
         only_short_term = max_delay == 0) %>% 
  count(only_short_term) %>% 
  ungroup() %>% 
  count(only_short_term)
```

To do so, we break down our dataset according to four types of study designs: experimental studies, quasi experimental studies, and observational studies. as presented in Table X, 
```{r}
dat %>% 
  select(unique_study_id, scale_type, study_design, d) %>% 
  filter(scale_type == "behavior") %>% 
  group_by(unique_study_id, study_design) %>% 
  summarise(mean_d_by_study_x_design = mean(d)) %>% 
  group_by(study_design) %>% 
  summarise(n = n(),
            mean_d_by_design = mean(mean_d_by_study_x_design)) %>% 
  kable(digits = 2, format = "latex", booktabs = T) %>%
  kableExtra::kable_styling() 
```

```{r}
dat |> 
  filter(scale_type == "attitudes") |> 
  distinct(unique_study_id, scale_name) |> 
  count(scale_name, sort = T) |> 
  View()
```

how many studies evaluate rape myths?
```{r rape_myth_outcome_prevalence}
dat %>%
  group_by(unique_study_id) %>%
  summarise(
    myth_var = any(str_detect(scale_name, "myth|IRMA|illinois"))
  ) %>%
  summarise(
    myth_count = sum(myth_var),
    no_myth_count = sum(!myth_var)
  )

```

what's the distribution of outcome types for bystander interventions

```{r bystander_dv_count}

# filter based on mentioning bystander in either title, description, 
# intervention name, or program name, or have a stated purpose of "4" (bystander)
bystander_dat <- dat %>%
  filter(
    str_detect(paper_title, "bystander") |
      str_detect(brief_description_of_the_intervention, "bystander") |
      str_detect(intervention_name, "bystander") |
      str_detect(program_name, "bystander") |
      str_detect(stated_purpose, "4")) %>%
    select(author, year, paper_title, unique_study_id, study_design,
           d, var_d, se_d, 
           behavior_type, has_both) %>%
  mutate(perp_vict_grouped = case_when(
    behavior_type == 'Victimization' | behavior_type == 'Perpetration' ~ 'perp_vict',
    behavior_type == 'Involvement' ~ 'involvement',
    behavior_type == 'Attitude' ~ 'Attitude',
    behavior_type == 'Bystander' ~ 'Bystander'))

bystander_dat %>%  study_count()

bystander_dat %>% group_by(behavior_type) %>% study_count()

bystander_dat %>% group_by(has_both) %>% study_count()

bystander_dat %>% filter(behavior_type != 'Involvement') %>% 
  split(.$perp_vict_grouped) %>% 
  map(.f = map_robust)

table(bystander_dat$behavior_type)    

# count studies that measure #just bystander behaviors
bystander_dat %>%
  filter(has_both != 'attitudes') %>% group_by(perp_vict_grouped) %>% study_count()

# double-check
bystander_dat_not_just_atts %>% 
  group_by(unique_study_id) %>%
  filter('Bystander' %in% behavior_type & !'perp_vict' %in% perp_vict_grouped) %>% ungroup() %>% study_count()


# what kind of effect sizes are we talking about? 
bystander_dat %>% filter(behavior_type != 'Involvement') %>% split(.$behavior_type) %>% map(map_robust)

# group perp and vict outcomes together

bystander_dat %>% filter(behavior_type != 'Involvement') %>% split(.$perp_vict_grouped) %>% map(map_robust)


# bystander interventions are a product of their time and place
bystander_dat %>% group_by(unique_study_id) %>% slice(1) %>% ungroup() %>%
  ggplot(aes(x = year)) +
  geom_line(stat = 'count', color = 'light blue') + 
  labs(x = "Year", 
       y = "Count", 
       title = "Bystander interventions over time") +
  # scale_x_date(date_breaks = "2 years") +
  scale_x_continuous(n.breaks = 20) +
  scale_y_continuous(n.breaks = 15) +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5))

## BTW an alternate way to do this
dat %>%
  group_by(unique_study_id) %>%
  summarise(
    bystander_count = ifelse(
      any(
        str_detect(c(
          brief_description_of_the_intervention, 
          intervention_name, program_name, paper_title), "bystander")) |
        any(str_detect(stated_purpose, "4")),
      TRUE, FALSE)
  ) %>% 
  summarise(
    bystander = sum(ifelse(is.na(bystander_count), FALSE, bystander_count)),
    no_bystander = sum(ifelse(is.na(bystander_count), TRUE, !bystander_count))
  )




```
